Create script.js
The JavaScript file identifies multiple variables including the question, button, and answer for the Magic Eight Ball page.  The options for the magic eight ball's responses can also be found here.  The event handler for the ask button allows for each click to generate a new response option from the list of output choices within the eight ball seen on the page.